Music, Scene Improvement, Sprites Realignment, Score
- Placeholder music during Menu and Gameplay
- Better sprite placement
- Preliminary Scenes with layout
- Playable game with serenade score

+++ Serenata/Assets/Data.cs
@@ -5,7 +5,7 @@ using UnityEngine.SceneManagement;
 
 public class Data : MonoBehaviour
 {
-    public float PlayerEndScore = 0.0f;
+    public int PlayerEndScore = 0;
 
     // Start is called before the first frame update
     void Start()
